Remove extra dependencies
