+++ packages/earljs/package.json
@@ -47,13 +47,10 @@
   "dependencies": {
     "debug": "^4.1.1",
     "jest-snapshot": "^26.6.2",
-    "lodash": "^4.17.15",
-    "pretty-format": "^26.6.2",
     "ts-essentials": "^6.0.5"
   },
   "devDependencies": {
     "@microsoft/tsdoc": "^0.13.0",
-    "@types/debug": "^4.1.5",
-    "@types/lodash": "^4.14.150"
+    "@types/debug": "^4.1.5"
   }
 }
