+++ packages/earljs/src/internals.ts
@@ -1,6 +1,4 @@
-export { Control } from './Control'
+export { Control, ValidationResult } from './Control'
 export { Expectation, ExpectationOptions, getControl } from './Expectation'
 export * from './matchers'
 export * from './plugins'
-export * from './validators/common'
-export * from './validators/smartEq'
